% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edabasic.R
\name{eda_variable}
\alias{eda_variable}
\title{Outlier Analysis & Correlation Analysis}
\usage{
eda_variable(data, variables = c(1:ncol(data)), log = FALSE)
}
\arguments{
\item{(data, variables=c(1:ncol(data)), log=FALSE)}{}
}
\value{

}
\description{
It does outlier bounds and count detection with and without log transformations. And correlation strength between continous variables.
}
